Convert species in setup_file into valid names
Subspecies names in the setup_file are now normalized before they
are interpreted as parameters. For example, the name "A-1" could
lead to problems when used to specialize a model parameter (as
lmfit.Parameter names are not allowed to contain '-'). Now, any
character that does not match a-zA-Zα-ωΑ-Ω0-9_ is converted into '_'.